All agents?

Report TRUE if all agents have their variable equal to a given value, report FALSE otherwise.

NLall(agents, world, var, val)

# S4 method for class 'matrix,worldMatrix,missing'
NLall(agents, world, val)

# S4 method for class 'matrix,worldArray,character'
NLall(agents, world, var, val)

# S4 method for class 'agentMatrix,missing,character'
NLall(agents, var, val)

Arguments

agents

Matrix (ncol = 2) with the first column pxcor and the second column pycor representing the patches coordinates, or

          `AgentMatrix` object representing the moving `agents`.

world

WorldMatrix or worldArray object.

var

Character. The name of the selected agents variable. If agents are patches and the world is a worldMatrix object, var must not be provided. If agents are patches and the world is a worldArray object, var is the name of the layer to use to define the patches values. If agents are turtles, var is one of the turtles' variable and can be equal to xcor, ycor, any of the variables created when turtles were created, as well as any variable created using turtlesOwn().

val

Numeric or character. Vector of any length.

Value

Logical. TRUE if all the agents have their variable equal to val, FALSE otherwise.

Details

world must not be provided if agents are turtles.

References

Wilensky, U. 1999. NetLogo. https://www.netlogo.org. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

See also

Author

Sarah Bauduin

Examples

# Patches
w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4, data = runif(25))
NLall(agents = patches(w1), world = w1, val = 5)
#> [1] FALSE
w2 <- w1
w2 <- NLset(world = w1, agents = patches(w1), val = 5)
NLall(agents = patches(w2), world = w2, val = 5)
#> [1] TRUE

# Turtles
t1 <- createTurtles(n = 5, coords = cbind(xcor = 1, ycor = 1), heading = c(1, 2, 2, 1, 2))
NLall(agents = t1, var = "xcor", val = 1)
#> [1] TRUE
NLall(agents = t1, var = "heading", val = 2)
#> [1] FALSE