The NetLogoR package

The suggested package fastshp can be installed with install.packages("fastshp", repos = "https://rforge.net", type = "source"). The examples included with the package, are located in the R package "examples" folder, which can be found at system.file(package = "NetLogoR", "examples").

Details

The 3 specific R examples can be opened here, but it's recommended to make copies of these for editing:

exPath <- file.path(system.file(package = "NetLogoR", "examples")
file.edit(exPath, "Ants", "Ants.R"))
file.edit(exPath, "Butterfly", "Butterfly-1.R"))
file.edit(exPath, "Wolf-Sheep-Predation", "Wolf-Sheep-Predation.R"))

See also

Author

Maintainer: Eliot J B McIntire eliot.mcintire@canada.ca (ORCID)

Authors:

Sarah Bauduin sarahbauduin@hotmail.fr (ORCID)

Alex M Chubaty achubaty@for-cast.ca (ORCID)

Other contributors:

His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]