Key base R functions for agentMatrix class

Slight modifications from the default versions.

# S4 method for class 'agentMatrix'
show(object)

# S4 method for class 'agentMatrix'
length(x)

# S4 method for class 'agentMatrix'
nrow(x)

# S3 method for class 'agentMatrix'
head(x, n = 6L, ...)

# S3 method for class 'agentMatrix'
tail(x, n = 6L, ...)

Arguments

object

An agentMatrix object.

x

An agentMatrix object

n

an integer vector of length up to dim(x) (or 1, for non-dimensioned objects).

...

arguments to be passed to or from other methods (currently, none used).

Value

show is called for its side effects. It shows all columns of data, except for the coordinates. To access those, use coordinates().

length returns a non-negative integer of length 1, except for vectors of more than 2^31 - 1 elements, when it returns a double.

nrow returns an integer of length 1 or NULL.

An agentMatrix object, like x, but generally smaller.