Patches here

Report the coordinates of the patches under the turtles locations.

patchHere(world, turtles)

# S4 method for worldNLR,agentMatrix
patchHere(world, turtles)

Arguments

world

WorldMatrix or worldArray object.

turtles

AgentMatrix object representing the moving agents.

Value

Matrix (ncol = 2) with the first column pxcor and the second column pycor representing the coordinates of the patches at the turtles

location. The order of the patches follows the order of the turtles.

Details

If a turtle is located outside of the world's extent, NA are returned for the patch coordinates.

References

Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

See also

Author

Sarah Bauduin

Examples

w1 <- createWorld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 10, coords = randomXYcor(w1, n = 10))
patchHere(world = w1, turtles = t1)
#>       pxcor pycor
#>  [1,]     2     6
#>  [2,]     3     1
#>  [3,]     4     0
#>  [4,]     9     7
#>  [5,]     1     1
#>  [6,]     5     9
#>  [7,]     9     1
#>  [8,]     7     6
#>  [9,]     2     2
#> [10,]     2     2