Patches on the left

Report the coordinates of the patches at the given distances of the turtles and given angle left of their headings.

patchLeft(world, turtles, dist, angle, torus = FALSE)

# S4 method for worldNLR,agentMatrix,numeric,numeric
patchLeft(world, turtles, dist, angle, torus = FALSE)

Arguments

world

WorldMatrix or worldArray object.

turtles

AgentMatrix object representing the moving agents.

dist

Numeric. Vector of distances from the turtles. dist must be of length 1 or of length turtles.

angle

Numeric. Vector of angles in degrees by which the turtle's headings should rotate to locate the patches. Must be of length 1 or of length turtles.

torus

Logical to determine if the world is wrapped. Default is torus = FALSE.

Value

Matrix (ncol = 2) with the first column pxcor and the second column pycor representing the coordinates of the patches at dist

distances of the turtles and angle to the left of their headings. The order of the patches follows the order of the turtles.

Details

If a given dist value is negative, then the turtle would look backward. If a given angle value is negative, then the turtle would look to the right.

     If `torus = FALSE` and the `patch` at distance `dist` of a `turtle`
     and `angle` degrees to the left of its `heading` is outside the
     `world`'s extent, `NA`
     are returned for the `patch` coordinates. If `torus = TRUE`, the `patch`
     coordinates from a wrapped `world` are returned.

References

Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

See also

Author

Sarah Bauduin

Examples

w1 <- createWorld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9)
t1 <- createTurtles(n = 1, coords = cbind(xcor = 2, ycor = 2), heading = 90)
patchLeft(world = w1, turtles = t1, dist = 2, angle = 90)
#>      pxcor pycor
#> [1,]     2     4