Rotate to the right

Rotate the turtles's headings to the right of angle degrees.

right(turtles, angle)

# S4 method for agentMatrix,numeric
right(turtles, angle)

Arguments

turtles
AgentMatrix object representing the moving agents.

angle
Numeric. Vector of angles in degrees by which to rotate the turtles' headings. Must be of length 1 or of length turtles.

Value

AgentMatrix representing the turtles with updated "heading" values.

Details

If a given angle value is negative, then the turtle rotates to the left.

References

Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

See also

https://ccl.northwestern.edu/netlogo/docs/dictionary.html#right

Examples

w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
t1 <- createTurtles(n = 10, world = w1)
of(agents = t1, var = "heading")

#>  [1] 122.89859 294.30133 166.50291 120.18549 244.78590 218.80100  73.36077
#>  [8] 109.13365 235.52906 342.84817

t1 <- right(turtles = t1, angle = 180)
of(agents = t1, var = "heading")

#>  [1] 302.89859 114.30133 346.50291 300.18549  64.78590  38.80100 253.36077
#>  [8] 289.13365  55.52906 162.84817