Subsetting and replacing for worldArray class

Subsetting and replacing for worldArray class

# S4 method for class 'worldArray,ANY,missing'
x[[i]]

# S4 method for class 'worldArray,ANY,missing'
x[[i]] <- value

# S4 method for class 'worldArray'
x$name

Arguments

x

A worldArray object.

i

Index number or layer name specifying a subset of layer(s) from the worldArray.

value

A replacement worldMatrix layer for one of the current layers in the worldArray.

name

Layer name, normally without back ticks, unless has symbols.

Value

The replacement method returns the original object, but with updated elements. The accessor method extracts the entire layer.

Examples

w1 <- createWorld(minPxcor = 0, maxPxcor = 9, minPycor = 0, maxPycor = 9, data = runif(100))
w2 <- createWorld(0, 9, 0, 9, data = runif(100))
w3 <- createWorld(0, 9, 0, 9, data = runif(100) + 2) # add 2 so different range
a1 <- stackWorlds(w1, w2)
a1[[2]]
#> class       : worldMatrix 
#> resolution  : 1, 1 (x, y)
#> dimensions  : Pxcor:  0 , 9 
#>               Pycor:  0 , 9 
#> names       :      layer 
#> min values  : 0.01146637 
#> max values  :  0.9863581 
#> First 4 rows and  4 columns:
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] 0.81715094 0.8500999 0.01507376 0.51384542
#> [2,] 0.29170274 0.2715587 0.98635813 0.08554531
#> [3,] 0.05158337 0.7084570 0.66163242 0.08839609
#> [4,] 0.61024855 0.5136914 0.86435378 0.85632395
a1[[2]] <- w3