Directions towards

Report the directions of each agents towards each corresponding agents2.

towards(agents, agents2, world, torus = FALSE)

# S4 method for matrix,matrix
towards(agents, agents2, world, torus = FALSE)

Arguments

agents

Matrix (ncol = 2) with the first column pxcor and the second column pycor representing the patches coordinates, or

          `AgentMatrix` object representing the moving `agents`.

agents2

Matrix (ncol = 2) with the first column pxcor and the second column pycor representing the patches coordinates, or

           `AgentMatrix` object representing the moving `agents`, or

           Matrix (`ncol` = 2) with the first column `x` and the second column
           `y` representing locations coordinates.

world

WorldMatrix or worldArray object.

torus

Logical to determine if the world is wrapped. Default is torus = FALSE.

Value

Numeric. Vector of angles in degrees of length equal to the largest number of agents/locations between agents and agents2.

Details

agents and agents2 must have the same number of agents/locations or if different, one of them must have only one agent/location. If agents and agents2 have the same number of agents/locations, the directions are calculated for each pair agents[i] and agents2[i] and not for each agents towards every single agents2.

     If `torus = FALSE`, `world` does not need to be provided.

     If `torus = TRUE` and the distance from one `agents` to
     its corresponding `agents2` is smaller around the
     sides of the `world` than across it, then the direction to `agents2`
     going around the sides of the `world` is returned.

     The direction from a patch to its location returns 0; the direction from
     a turtle to its location returns the turtle's heading.

References

Wilensky, U. 1999. NetLogo. http://ccl.northwestern.edu/netlogo/. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

See also

Author

Sarah Bauduin

Examples

w1 <- createWorld(minPxcor = 0, maxPxcor = 4, minPycor = 0, maxPycor = 4)
towards(agents = patches(w1), agents2 = cbind(x = 0, y = 0))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'agents' in selecting a method for function 'towards': unable to find an inherited method for function 'patches' for signature '"worldMatrix"'
t1 <- createTurtles(n = 10, world = w1)
towards(agents = t1, agents2 = cbind(x = 0, y = 0))
#>  [1] 225 225 225 225 225 225 225 225 225 225